pyproject.toml
[tool.poetry]
name = "personal-analyzer-bot"
version = "0.1.0"
description = ""
authors = ["Maksim <maksim@devopalev.ru>"]
readme = "README.md"
packages = [{include = "personal_analyzer_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
sqlalchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
alembic = "^1.11.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.19.0"